requirements.txt
flask>=2.0.0
flask-sqlalchemy>=2.5.0
flask-migrate>=3.0.0
flask-cors>=3.0.0
python-dotenv>=0.19.0
web3>=5.24.0
py-solc-x>=1.1.0
eth-account>=0.5.7
eth-typing>=2.2.2
pyjwt>=2.3.0
python-jose>=3.3.0
requests>=2.26.0
werkzeug>=2.0.0 